src/test/kotlin/DebugHandlerTest.kt
import org.junit.jupiter.api.Assertions.*
import org.junit.jupiter.api.Test

class DebugHandlerTest {
    @Test
    fun testLoadPath() {
        val debugHandler = DebugHandler("debugger/path")
        debugHandler.loadPath("/path/to/debugger")
        assertEquals("/path/to/debugger", debugHandler.path)
    }

    @Test
    fun testSetBreakpoints() {
        val debugHandler = DebugHandler("debugger/path")
        debugHandler.setBreakpoints("testFile", listOf(10, 20, 30))
        assertEquals("testFile", debugHandler.fileName)
        assertEquals(listOf(10, 20, 30), debugHandler.breakPoints)
    }
    // Add more test cases for other functions as needed

    @Test
    fun testSetBreakHandler() {
        val debugHandler = DebugHandler("debugger/path")
        var called = false
        debugHandler.setBreakHandler { called = true }
        debugHandler.breakHandlerFun.invoke()
        assertEquals(true, called)
    }

    @Test
    fun testResume() {
        val debugHandler = DebugHandler("debugger/path")
        debugHandler.resume()
        assertEquals(false, debugHandler.paused)
    }

    fun testResumeExecution() {
        val debugHandler = DebugHandler("debugger/path")
        debugHandler.resumeExecution()
        // Hard to test this method as it only writes to the stream
        // We can verify if the method executes without error
    }

    @Test
    fun testGetBackTrace() {
        val debugHandler = DebugHandler("debugger/path")
        // Testing with a mock BufferedReader and BufferedWriter could be more comprehensive
        // Here, we can only test the basic behavior
        val backTrace = debugHandler.getBackTrace()
        assertEquals("Backtrace was not found", backTrace)
    }
}
